+++ claude_tap/cli.py
@@ -486,6 +486,7 @@ async def _async_main(args: argparse.Namespace) -> int:
                     trace_methods=CLIENT_CONFIGS[args.client].forward_trace_methods,
                     trace_path_prefixes=CLIENT_CONFIGS[args.client].forward_trace_path_prefixes,
                     trace_path_suffixes=CLIENT_CONFIGS[args.client].forward_trace_path_suffixes,
+                    trace_model_requests_only=CLIENT_CONFIGS[args.client].forward_trace_model_requests_only,
                     store_stream_events=args.store_stream_events,
                     capture_only=capture_only,
                 )